rename test type 'OTHER' to 'PROGRAM'
