+++ src/main/java/fr/kungfunantes/backend/model/test/program/ProgramTest.java
@@ -1,4 +1,4 @@
-package fr.kungfunantes.backend.model.test.other;
+package fr.kungfunantes.backend.model.test.program;
 
 import fr.kungfunantes.backend.model.exercise.ExerciseScale;
 import fr.kungfunantes.backend.model.test.Test;
@@ -13,8 +13,8 @@ import java.util.List;
 @Data
 @Entity
 @ApiModel
-@DiscriminatorValue(value = "OTHER")
-public class OtherTest extends Test {
+@DiscriminatorValue(value = "PROGRAM")
+public class ProgramTest extends Test {
     @OneToMany(cascade = CascadeType.ALL)
     @OrderBy("position ASC")
     private List<ExerciseScale> exercisesScales;
